GEODE-3314: Fix DLockService token leak.
And add a test so that we don't do that again.

We had been unlocking the DLock on the remote server before freeing the
local `DLockToken`. This meant that the local token got reused by
another thread acquiring the same lock and the count went up and up and
up (due to some logic that wouldn't decrease the count if it no longer
owned the token).

Tokens should no longer have usageCounts that climb over time, which
means they should be released normally by freeResources.

Signed-off-by: Bruce Schuchardt <bschuchardt@pivotal.io>